requirements.txt
python-dotenv
slack-sdk
certifi
requests
